Refactor of tests: put in appropriate packages
